Backend API/requirement.txt
requests==2.22.0
marshmallow==3.3.0
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.4.1
Flask-Marshmallow==0.10.1
marshmallow-sqlalchemy==0.21.0
Flask-Uploads==0.2.1
Flask-Cors==3.0.8
python-dotenv==0.10.3
werkzeug==0.16.0
tensorflow==1.14.0
numpy==1.16.5
pillow==6.2.0
uwsgi
psycopg2
